# specifications/openapi.yaml
...
openapi: 3.0.2
info:
  description: >
    A self guided random access training system on subjects related to identifying
    and evaluating opportunities to improve industrial energy, 
    resource and production efficiency.

  version: 1.0.0
  title: EEC Walkthrough API
  contact:
    name: API support
    email: 'thomasza@oregonstate.edu'
...
paths:
  /users:
    post:
      summary: Creates a new user
      description: >
        Creates a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userName
                - firstName
                - lastName
                - password
                - email
              properties:
                userName:
                  example: User2020
                password:
                  example: aseoe23oexZw23!
                firstName:
                  example: John
                lastName:
                  example: Doe
                email:
                  example: 'doejohn@oregonstate.edu'
      responses:
        '201':
          $ref: '#/components/responses/User_POST_201'
        '400':
          $ref: '#/components/responses/400'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  '/users/login/{userName}/{password}':
    get:
      summary: Login the user
      description: >
        Verifies the user's credentials
      operationId: validateUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/User_userName'
        - $ref: '#/components/parameters/User_password'
      responses:
        '200':
          $ref: '#/components/responses/User_GET_200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/users/{userId}':
    get:
      summary: Gets a user by ID
      description: >
        Gets a user by ID
      operationId: getUser
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/User_userId'
      responses:
        '200':
          $ref: '#/components/responses/User_GET_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  
    patch:
      summary: Update user info
      description: >
        Updates the info about the user
      operationId: patchUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  example: User2020
                password:
                  example: aseoe23oexZw23!
                firstName:
                  example: John
                lastName:
                  example: Doe
                email:
                  example: 'doejohn@oregonstate.edu'
                role:
                  example: 1
      parameters:
        - $ref: '#/components/parameters/User_userId'
      responses:
        '200':
          $ref: '#/components/responses/PATCH_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  /pages:
    post:
      summary: Creates a new page
      description: >
        Creates a new subject or industry page
      operationId: createPage
      tags:
        - Pages
      requestBody:
        $ref: '#/components/requestBodies/Page_POST'
      responses:
        '201':
          $ref: '#/components/responses/Page_POST_201'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  '/pages/industries/{industryId}/subjects/{subjectId}':
    post:
      summary: Add a subject to an industry
      description: >
        Add a subject to an industry
      operationId: createIndustrySubject
      tags:
        - Pages
      parameters:
        - in: path
          name: industryId
          schema:
            type: integer
          required: true
          description: The ID of the industry page
        - in: path
          name: subjectId
          schema:
            type: integer
          required: true
          description: The ID of the subject page
      responses:
        '201':
          $ref: '#/components/responses/Page_POST_MANY_201'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'

    delete:
      summary: Remove a subject from an industry
      description: >
        Remove a subject from an industry
      operationId: deleteIndustrySubject
      tags:
        - Pages
      parameters:
        - in: path
          name: industryId
          schema:
            type: integer
          required: true
          description: The ID of the industry page
        - in: path
          name: subjectId
          schema:
            type: integer
          required: true
          description: The ID of the subject page
      responses:
        '200':
          $ref: '#/components/responses/DELETE_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/pages/{pageId}/all':
    get:
      summary: Get all of the content contained in one page
      description: >
        Get all of the content contained in one page.
        This includes page info, headers, cards, and items
      operationId: getPageAll
      tags:
        - Pages
      parameters:
        - $ref: '#/components/parameters/Page_pageId'
      responses:
        '200':
          $ref: '#/components/responses/Page_GET_ALL_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/pages/all':
    get:
      summary: Get a list of all pages and their related subjects/industries
      description: >
        Get the id and name of every page as well as an array of
        the related subjects/industries.
      operationId: getPagesAll
      tags:
        - Pages
      responses:
        '200':
          $ref: '#/components/responses/Page_GET_LIST_ALL_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/pages/{pageId}':
    get:
      summary: Gets a page by ID
      description: >
        Gets a subject or industry page by ID
      operationId: getPage
      tags:
        - Pages
      parameters:
        - $ref: '#/components/parameters/Page_pageId'
      responses:
        '200':
          $ref: '#/components/responses/Page_GET_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      summary: Update a page by ID
      description: >
        Update a subject or industry page by ID
      operationId: updatePage
      tags:
        - Pages
      requestBody:
        $ref: '#/components/requestBodies/Page_PATCH'
      parameters:
        - $ref: '#/components/parameters/Page_pageId'
      responses:
        '200':
          $ref: '#/components/responses/PATCH_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'

    delete:
      summary: Delete a page by ID
      description: >
        Delete a subject or industry page by ID
      operationId: deletePage
      tags:
        - Pages
      parameters:
        - $ref: '#/components/parameters/Page_pageId'
      responses:
        '200':
          $ref: '#/components/responses/DELETE_200'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
...
components:
  schemas:
...
    Id:
      type: integer
      format: int32
      minimum: 1
      maximum: 4294967295
      example: 15
...
    UtcTime:
      type: string
      format: 'yyyy-mm-dd HH:MM:SS'
      minLength: 19
      maxLength: 19
      example: '2020-12-31 23:59:59'
...
  requestBodies:
...
    Page_POST:
      required: true
      content:
        application/json:
          schema:
            properties:
                    pageType:
                      example: 1
                    name:
                      example: Compressed Air
                    title:
                      example: Compressed air is a common utility found in most industrial facilities
                    description:
                      example: Compressed air has been a key industrial utility since the...
                    imageUrl:
                      example: http://flip2.engr.oregonstate.edu:1111/images/aircompressor.png
                    userId:
                      allOf:
                        - $ref: '#/components/schemas/Id'
    Page_PATCH:
      required: true
      content:
        application/json:
          schema:
            properties:
                    pageType:
                      example: 1
                    name:
                      example: Compressed Air
                    title:
                      example: Compressed air is a common utility found in most industrial facilities
                    description:
                      example: Compressed air has been a key industrial utility since the...
                    imageUrl:
                      example: http://flip2.engr.oregonstate.edu:1111/images/aircompressor.png
                    approved:
                      example: 1
...
  parameters:
    User_userId:
      name: userId
      in: path
      description: The unique identifier of the user
      required: true
      schema:
        $ref: '#/components/schemas/User/properties/userId'
...
    User_userName:
      name: userName
      in: path
      description: The unique account name of the user
      required: true
      schema:
        $ref: '#/components/schemas/User/properties/userName'
    User_password:
      name: password
      in: path
      description: The password associated with the user's account
      required: true
      schema:
        $ref: '#/components/schemas/User/properties/password'
    Page_pageId:
      name: pageId
      in: path
      description: The unique identifier of the page
      required: true
      schema:
        $ref: '#/components/schemas/Page/properties/pageId'
...
  responses:
    User_POST_201:
      description: User created
      content:
        application/json:
          schema:
            description: An object containing information about the new user
            type: object
            required:
              - Id
            properties:
              insertId:
                $ref: '#/components/schemas/Id'
...
    User_GET_200:
      description: An object containing information about the user
      content:
        application/json:
          schema:
            properties:
              userId:
                example: 20
              userName:
                example: User2020
              firstName:
                example: John
              lastName:
                example: Doe
              email:
                example: 'doejohn@oregonstate.edu'
              role:
                example: 1
    PATCH_200:
      description: An object containing information about the changed rows
      content:
        application/json:
          schema:
            properties:
              changedRows:
                example: 3
...
    Page_POST_201:
      description: Page created
      content:
        application/json:
          schema:
            description: An object containing information about the new page
            type: object
            required:
              - Id
            properties:
              insertId:
                $ref: '#/components/schemas/Id'
    Page_POST_MANY_201:
      description: Page created
      content:
        application/json:
          schema:
            description: An object containing information about the industry / subject relationship
            type: object
            required:
              - Id
            properties:
              industryId:
                example: 10
              subjectId:
                example: 15
    Page_GET_200:
      description: An object containing information about the page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Page'
    Page_GET_LIST_ALL_200:
      description: An object containing information about all pages seperated into subjects and industries
      content:
        application/json:
          schema:
            properties:
              pages:
                example: {
                  "subjects": [
                {
                   "pageId": 5,
                   "pageType": 0,
                   "name:": "Refrigeration",
                   "relatedPages": [ 
                     {
                       "pageId": 10,
                       "name": "Electricity"
                     }
                    ]
                }
              ],
                "industries": [
                {
                   "pageId": 10,
                   "pageType": 1,
                   "name:": "Electricity",
                   "relatedPages": [ 
                     {
                       "pageId": 5,
                       "name": "Refrigeration"
                     }
                    ]
                }
              ]
            }
    Page_GET_ALL_200:
      description: An object containing information about the page and all of its content
      content:
        application/json:
          schema:
            properties:
              pageId:
                allOf:
                  - description: The unique identifier of the page
                  - $ref: '#/components/schemas/Id'
              pageType:
                description: Value indicating if the page is a subject page (0) or an industry page (1)
                type: boolean
                minLength: 0
                maxLength: 1
                example: 1
              name:
                description: The name of the page
                type: string
                minLength: 1
                maxLength: 100
                example: Compressed Air
              title:
                description: A title for the paragrah uses to present the page topic
                type: string
                minLength: 1
                maximum: 100
                example: Compressed air is a common utility found in most industrial facilities
              description:
                description: A short description of the page topic
                type: string
                minLength: 1
                maxLength: 1000
                example: Compressed air has been a key industrial utility since the...
              imageUrl:
                description: The image to accompany the short description of the page topic
                type: string
                minLength: 1
                maxLength: 500
                example: http://flip2.engr.oregonstate.edu:1111/images/aircompressor.png
              userId:
                allOf:
                  - description: The user ID of the user who created the page
                  - $ref: '#/components/schemas/Id'
              created:
                allOf:
                  - description: The time that the page was created
                  - $ref: '#/components/schemas/UtcTime'
              approved:
                description: Value indicating if the page has been approved to be shown on the website
                type: boolean
                minimum: 0
                maximum: 1
                example: 0
              headers:
                example: [ 
                 {
                   "headerId": 15,
                   "pageId": 15,
                   "title": "Compressed Air Opportunities to Consider",
                   "userId": 15,
                   "created": "2020-12-31 23:59:59",
                   "approved": 0,
                   cards: [{
                           cardId: 15,
                           headerId: 15,
                           title: Best Practices,
                           userId: 15,
                           created: '2020-12-31 23:59:59',
                           approved: 0,
                           items: [{
                                    itemId: 15,
                                    cardId: 15,
                                    parentId: 15,
                                    iconType: 5,
                                    contentText: This guide focuses mainly on screw and reciprocating compressors. These are the most common...,
                                    contentUrl: 'https://www.compressedairchallenge.org/data/sites/1/media/library/sourcebook/Improving_Compressed_Air-Sourcebook.pdf',
                                    contentLabel: Assessing Industrial Air Compressor,
                                    userId: 15,
                                    created: '2020-12-31 23:59:59',
                                    approved: 0
                                  }]
                          }]
                 }
                ]
    DELETE_200:
      description: The number of rows deleted
      content:
        application/json:
          schema:
            description: An object containing the number of rows deleted
            type: object
            required:
              - Id
            properties:
              affectedRows:
                type: integer
                example: 1
...
    '400':
      description: Missing or invalid request body
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Bad request.
    '401':
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Unauthorized.
    '403':
      description: Not allowed to perform this action
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Forbidden.
    '404':
      description: Specified resource not found
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Not found.
    '422':
      description: Unprocessable entity
      content:
        application/json:
          schema:
            properties:
              errors:
                description: A list of input validation errors
                type: array
                example: [
                      {
                          "value": "",
                          "msg": "Invalid value",
                          "param": "userName",
                          "location": "body"
                      },
                      {
                          "value": "",
                          "msg": "Invalid value",
                          "param": "email",
                          "location": "body"
                      }
                  ]
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: An internal server error occurred. Please try again later.

# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/silverware13/EEC_Walkthrough/1.0.0